Implement direct mode

This closes #27

+++ test_scripts/build_with_buildcache.sh
@@ -1,19 +1,22 @@
 #!/bin/bash
+# These shellcheck warnings are handled implicitly by trap_handler:
+# shellcheck disable=SC2164
 
-# Check if we're running on Windows (this is mostly a trick to get the script
-# working on the Travis-CI Windows slaves that run Git bash).
+# Check if we're running on Windows.
 function is_windows() {
-    if [ "$(expr substr $(uname -s) 1 5)" == "MINGW" ]; then
+    local u
+    u=$(uname -s)
+    if [ "${u:0:5}" == "MINGW" ]; then
         return 0
     fi
     return 1
 }
 
 function cleanup() {
-    [[ ! -z "${PROJDIR}" ]] && rm -rf "${PROJDIR}"
-    [[ ! -z "${BUILDCACHE_DIR}" ]] && rm -rf "${BUILDCACHE_DIR}"
-    [[ ! -z "${LOG_DIR}" ]] && rm -rf "${LOG_DIR}"
-    [[ ! -z "${SYMLINKSDIR}" ]] && rm -rf "${SYMLINKSDIR}"
+    [[ -n "${PROJDIR}" ]] && rm -rf "${PROJDIR}"
+    [[ -n "${BUILDCACHE_DIR}" ]] && rm -rf "${BUILDCACHE_DIR}"
+    [[ -n "${LOG_DIR}" ]] && rm -rf "${LOG_DIR}"
+    [[ -n "${SYMLINKSDIR}" ]] && rm -rf "${SYMLINKSDIR}"
 }
 
 # Treat errors with exit code 1.
@@ -34,12 +37,17 @@ if is_windows; then
     EXESUFFIX=".exe"
 fi
 
-# Note: This script is expected to be run from the BuildCache build folder.
+# Find the buildcache executable.
 BUILDCACHEDIR="$(pwd)"
+BUILDCACHEEXE="${BUILDCACHEDIR}/buildcache${EXESUFFIX}"
+if [[ ! -x "${BUILDCACHEEXE}" ]]; then
+    echo "ERROR: Not a BuildCache executable: ${BUILDCACHEEXE}"
+    echo "Note: This script is expected to be run from the BuildCache build folder."
+    exit 1
+fi
+echo "BuildCache executable: ${BUILDCACHEEXE}"
 
 # Install BuildCache and symlinks to BuildCache for common compilers.
-BUILDCACHEEXE="${BUILDCACHEDIR}/buildcache${EXESUFFIX}"
-echo "Buildcache executable: ${BUILDCACHEEXE}"
 SYMLINKSDIR="${BUILDCACHEDIR}/symlinks"
 echo "Install symlinks in: ${SYMLINKSDIR}"
 rm -rf "${SYMLINKSDIR}" ; mkdir -p "${SYMLINKSDIR}"
@@ -56,6 +64,7 @@ export PATH="${SYMLINKSDIR}:${BUILDCACHEDIR}:${PATH}"
 export BUILDCACHE_DIR=/tmp/.buildcache-$$
 rm -rf "${BUILDCACHE_DIR}" ; mkdir -p "${BUILDCACHE_DIR}"
 export BUILDCACHE_DEBUG=2
+export BUILDCACHE_DIRECT_MODE=true
 export LOG_DIR=/tmp/.buildcache-logs-$$
 rm -rf "${LOG_DIR}" ; mkdir -p "${LOG_DIR}"
 
@@ -81,27 +90,46 @@ buildcache -s
 echo "--- LOG (pass 1) ---"
 cat "${LOG_DIR}/test-pass-1.log"
 
-echo "======== Second build (warm cache) ========"
+echo "======== Second build (warm cache, direct mode OFF ) ========"
 cd "${PROJDIR}"
 rm -rf "${BUILDDIR}" ; mkdir -p "${BUILDDIR}"
 cd "${BUILDDIR}"
 cmake -G Ninja -DCMAKE_BUILD_TYPE=Release "${SRCDIR}"
 
-time BUILDCACHE_LOG_FILE="${LOG_DIR}/test-pass-2.log" ninja
+time BUILDCACHE_LOG_FILE="${LOG_DIR}/test-pass-2.log" BUILDCACHE_DIRECT_MODE=false ninja
 buildcache -s
 echo "--- LOG (pass 2) ---"
 cat "${LOG_DIR}/test-pass-2.log"
+
+echo "======== Third build (warm cache, direct mode ON ) ========"
+cd "${PROJDIR}"
+rm -rf "${BUILDDIR}" ; mkdir -p "${BUILDDIR}"
+cd "${BUILDDIR}"
+cmake -G Ninja -DCMAKE_BUILD_TYPE=Release "${SRCDIR}"
+
+time BUILDCACHE_LOG_FILE="${LOG_DIR}/test-pass-3.log" ninja
+buildcache -s
+echo "--- LOG (pass 3) ---"
+cat "${LOG_DIR}/test-pass-3.log"
 
 # Make sure that there were no cache misses during the second pass.
 EXIT_CODE=0
 if grep -q "Cache miss" "${LOG_DIR}/test-pass-2.log"; then
     echo "*** FAIL: The second build pass contained cache misses."
     EXIT_CODE=1
-else
-    echo "The test passed!"
+fi
+
+# Make sure that there were no direct mode cache misses during the third pass.
+if grep -q "Direct mode cache miss" "${LOG_DIR}/test-pass-3.log"; then
+    echo "*** FAIL: The third build pass contained cache misses."
+    EXIT_CODE=1
 fi
 
 # Clean up.
 cleanup
 
+if [[ "${EXIT_CODE}" == "0" ]]; then
+    echo "----------------------------------------------------------"
+    echo "The test passed!"
+fi
 exit ${EXIT_CODE}
